Wifi: cw1200: Allow build as a module.
The 'cw1200_spi' is a seperate module from 'cw1200_core'.  The power
management routines are part of the 'core'.  We put some power management
stubs in 'cw1200_spi'.  This doesn't work when linking as a module.  The
'cw1200_core' must stand alone.  Previously, we broke builds with sdio as
well.  With this change, all pm stubs are inlined so the compiler will just
eliminate them.   This generates better code and allows more build variations.