cw1200: Statically link firmware to the module/code.
The firmware is always present with the code so that driver
restarts will have firmware available on suspend/resume, where
the system disk maybe in-accessible.

As it is linked with the code, when it is a module and not loaded
other platforms have no memory penalty.  There are also fewer license
issues as non-GPL code has not been linked directly to the kernel when
it is a module.