net:wireless:cw1200: Fix use-after-free in cw1200_unregister_common().
The 'cw1200_common' structure is allocated by ieee80211_alloc_hw() in
the function cw1200_init_common().  The 'cw1200_common' (often called
'priv') is allocated along with a ieee80211_hw struct and an 'rdev'
network structure (it is struct cfg80211_registered_device).  The main
'rdev' is a reference counted structure.

Depending on outstanding network activity, the reference count maybe non
zero when 'cw1200_unregister_common()' is called.  If we are unfortunate,
and the reference is zero, then after ieee80211_unregister_hw(), the
'priv' pointer is no longer valid.  This is most easily seen with DEBUG_SLAB,
which will poison freed memory.  However, various bad things can happen
(infinite loops, etc) when ever we have 'use-after-free' errors.

For NBS, this is especially critical as we do 'firmware reset' where we
restart the driver.  Otherwise, this is mainly only used during module
unloading.