cw1200: Convert BUG_ON() to CW1200_BUG_ON().
The new macro prints a message, does a stack dump and turns off the
device.  If the user powers on, a remote diagnostic upload should
give the 'BUG()' location and a stack trace.

Previously, the 2.6.36 version would just 'busy loop' and only a
person with a console connected and verbose debug can see any of
the messages.