cw1200: Use completion instead of workqueue + variable.
The 'hw_bufs_used' was a flag that indicated that a REQ/RESP message
had been sent to the cw1200 chip.  However, there is no restriction
on using multiple buffers.  If multiple buffers are used, then the
previous code says that is an error.  However, back-to-back cw1200
commands maybe in progress.  It is not an error.  Just use a completion
to signal a response.  The 'wsm_cmd_mux' mutex is already preventing
multiple c1200 commands at once.