cw1200: sta: Keep 'conf_mutex' during info_changed and fix join complete.
We keep the 'conf_mutex' for the whole bss_info_changed as we are
doing 'unjoin' and starting 'join' work.  This prevents other 'jobs' from
starting.  The bss_info_changed() is called as an ioctl() from wpa_supplicant.

Also, always schedule a 'join complete', however, abort the join complete
work if there is no 'join_timeout'; this means that the join has been
cancelled already.