cw1200:bh: Reduce race and improve debug.
Print out the current tick and lock status when 'firmware locks'.
Also, the main 'bh' wake condition has a race.  This code,

	rx = atomic_xchg(&priv->bh_rx, 0);  /* ???? */
	tx = atomic_xchg(&priv->bh_tx, 0);  /* ???? */
	term = atomic_xchg(&priv->bh_term, 0); /* ??? */
	suspend = pending_tx ? 0 : atomic_read(&priv->bh_suspend); /* ??? */
	(rx || tx || term || suspend || priv->bh_error);

The lines with '???' may be interrupted and rescheduled.  So some other thread
may change 'bh_rx' and 'bh_tx' after they are read.  This could cause the bottom
half to stay asleep for a prolonged time.