cw1200: Fix/triage race between bottom half and other contexts.
This is a structural issue with this driver.  This code is an attempt
to patch the issue.  We have network packets and '802.11' messages called
'wsm' by the driver.  The 'wsm' message may 're-associate' etc.  Obviously,
the cw1200 chip can not handle network (IP) traffic, while it is migrating
AP/SSID/BSSID, etc.  The driver tries to handle this with 'tx_lock'.  However,
the bottom half may read 'tx_lock' and then proceed to remove a queued IP packet.

This IP packet is sent and the 'wsm' thread will think that we are ready to send
management packets, such as the 'reset' command.   Also the driver has a limit of
16 buffers.  The variable 'hw_bufs_used' keeps track of this.  It includes both
'IP' and 'wsm' packets/messages.  Replies are returned and the driver decrements
the count.  However, if we send an IP and then a 'reset', the driver never returns
the sucess/fail of the original IP packet and then 'hw_bufs_used' is never decremented.

This locks the driver and it prints out 'firmware stuck' and/or 'missed interrupt'.
Neither of these is the case and it is a race within this driver.