cw1200: Remove wsm_lock_tx_async(), moving flush to main workqueues.
Various places would do an async() lock and then have the main work
do the flush.  Now, we just call the 'wsm_lock_tx()' in the workqueue,
which will do the flush if it is first.

As well, take the 'wsm_cmd' mutex in the lock to stop 2nd callers from
issuing a command while the flush is in progress.