cw1200: Implement hook to supply 'wlan0/wireless/link' value.
This allows the application group to do nothing.  It is what
they are good at.

The cw1200 chip records a recieved signal with every received
packet, now copied to the driver structure.  The value is
converted from the cw1200 space to a range that 'link' normally
prints.  Ie, dB to some 0-70 value, which is magic according to
some app developers.